Added auto-calculation of timeouts
It's very basic right now. When it has received enough samples,
it will calculate a threshold and set it.

The very simple calculation (with no validation right now) is
to take the median of the time between beats and add an additional
50 percent.

So if beats are coming with 10 seconds between them, the timeout
will be set to 15 seconds.