New 'snapshot' file backend
Previously, services and views were in one file per service/view
and they were rewritten whenever they were updated (for every beat,
that is). This resulted in a lot of files and quite many I/O
operations when having 100+ services. Also, backups were a bit
problematic due to the large number of files.

The new file backend saves a snapshot of services/views into one
compressed file every 60 seconds. This is done in the background.