Install go-bindata before we try to use it.
This reduces the number of manual steps necessary to build
lovebeat at the reasonable cost of a most of the time no-op
"go install" run when an asset source file changes.

We could've saved those cycles by instead doing

    dashboard/assets.go: $(ASSETS) | $(GO_BINDATA)
            $(GO_BINDATA) $(BINDATA_DEBUG) ...

    $(GO_BINDATA):
            go install github.com/jteeuwen/go-bindata/go-bindata

but then a build following "make clean" wouldn't have refreshed
the go-bindata binary even if it was out of date. Changing the
go-bindata source won't trigger a rebuild of assets.go, though.