Preparation refactoring
We're moving persistent stuff to the configuration files and
letting the 'zero configuration' idea cover services - not views,
as the views take care of monitoring and alerting and we want that
to be easily handled by provisioning tools that would rather want
to work with configuration files than a REST interface.

That was a long sentence.