Generate version number automatically with "git describe".
Instead of hardcoding the version number in a constant in
lovebeat.go we extract the constant to version.go and generate
that file in every build. The version number now comes from
the output of "git describe", which describes the currently
checked out commit in an unambiguous manner, taking the closest
tag into account. The point of this is that we can now check out
any commit we like and get a correct and useful version number.

We've chosen to not ship a default version.go with the Lovebeat
source code, hence forcing people to have make(1) generate the file.
This means that make(1) must be used at least once to be able to
build Lovebeat. The build instructions have been updated to reflect
this. This unfortunately also means that all users will have to
rebuild the dashboard assets (unless they hack the makefile to
remove that dependency).