Automatically update debian/changelog with the current version.
Along the lines of automatically updating version.go with a current
and useful version number to compile into the binary we want the
same to apply to the Debian packages being built; it should be
possible to check out any commit and build a Debian package from it.

We accomplish this by checking if the current version is different
from the most recent entry in debian/changelog, used by the Debian
build system when creating a package. If necessary we create a new
entry with the current version.

This will in most cases create a dirty debian/changelog every time
a package is built but that's a better problem to have.

We also change the current debian/changelog entry from UNRELEASED
to RELEASED so that dch(1) will create a new entry instead of
replacing the old one.