KISS: One timeout per service only
NOTE: This is a backwards incompatible change on our way to 1.0.

Warnings and errors were confusing, especially when you connect
an alert to a view. Should it trigger for warnings or errors, or
both?

Now we let the user decide. If the user wants different levels,
he can add this information to the service name and have dynamic
views filtering on each type and having alerts trigger in any
way they want.