Rate limit beats
Lovebeat will easily beats coming in at a high frequency, but since
they are propagated to the dashboard via the websocket bus, the
browser may even freeze due to the large amount of updates it has to
do in the DOM tree.

So limit this problem in the source by rate limiting beats, allowing
them to come no faster than once per second.